+++ src/pages/dashboard/calculator/index.tsx
@@ -1,33 +1,15 @@
-import { initializeApp } from 'firebase/app';
-import { collection, getDocs, getFirestore, query } from 'firebase/firestore';
-import { useCallback, useEffect, useState } from 'react';
-import { FIREBASE_API } from 'src/config-global';
-import DashboardLayout from 'src/layouts/dashboard/DashboardLayout';
+import { useRouter } from 'next/router';
+import { useEffect } from 'react';
+import { PATH_DASHBOARD } from 'src/routes/paths';
 
 const Page = () => {
-  const [calculations, setCalculations] = useState<any[]>([]);
-  const firebaseApp = initializeApp(FIREBASE_API);
-  const DB = getFirestore(firebaseApp);
-
-  const getCalculations = useCallback(async () => {
-    const q = query(collection(DB, 'importCalculations'));
-    const querySnapshot = await getDocs(q);
-    querySnapshot.forEach((doc) => {
-      setCalculations((prev) => [...prev, doc.data()]);
-    });
-  }, [DB]);
+  const { push } = useRouter();
 
   useEffect(() => {
-    getCalculations();
-  }, [getCalculations]);
+    push(PATH_DASHBOARD.scientech.calculator.list);
+  });
 
-  return (
-    <DashboardLayout>
-      <div>
-        <pre>{JSON.stringify(calculations, null, 2)}</pre>
-      </div>
-    </DashboardLayout>
-  );
+  return null;
 };
 
 export default Page;
