+++ src/routes/paths.ts
@@ -41,6 +41,14 @@ export const PATH_DASHBOARD = {
   fileManager: path(ROOTS_DASHBOARD, '/files-manager'),
   permissionDenied: path(ROOTS_DASHBOARD, '/permission-denied'),
   blank: path(ROOTS_DASHBOARD, '/blank'),
+  scientech: {
+    home: path(ROOTS_DASHBOARD, '/home'),
+    calculator: {
+      list: path(ROOTS_DASHBOARD, '/calculator/list'),
+      new: path(ROOTS_DASHBOARD, '/calculator/new'),
+      view: (id: string) => path(ROOTS_DASHBOARD, `/calculator/${id}`),
+    },
+  },
   general: {
     app: path(ROOTS_DASHBOARD, '/app'),
     ecommerce: path(ROOTS_DASHBOARD, '/ecommerce'),
